Move Token struct under persistent_tokens
