--- as/cNFT/assembly/models/token.ts
@@ -1,10 +1,0 @@
-@nearBindgen
-export class Token {
-    id: string
-    owner_id: string
-    creator: string
-    prev_owner: string
-    next_approval_id: number
-    media: string
-    extra: string
-}
